HttpApi: Add doc
